Moving the largest data sets to google drive
Its generally good to keep the size of a github repo under 10MB if
possible.  This means putting large data sources into a different place
if possible.  Google drive is a good choice because contents are easily
made public.  Readme.md has been updated with a link to the new
location for these.